f5
Former-commit-id: c3c20fee02acb1c9eec7332ec173e05e9df9fb99 [formerly a0bdb02a4ec4299e767d011e86970412be10d25c]
Former-commit-id: 495e626a762ffa4d8df3f66dd6e3d7c896950f94

+++ src/utils/dataset.py
@@ -19,11 +19,11 @@ class SphereDataset(Dataset):
         img_path = os.path.join(self.image_dir, self.images[index])
         image = np.array(Image.open(img_path).convert("RGB"), dtype=np.uint8)
 
-        mask = np.zeros((image.shape[0], image.shape[1]), dtype=np.float32)
+        mask = np.zeros((image.shape[0], image.shape[1]), dtype=np.uint8)
 
         if self.transform is not None:
             augmentations = self.transform(image=image, mask=mask)
             image = augmentations["image"]
-            mask = augmentations["mask"]
+            mask = augmentations["mask"].float()
 
         return image, mask
